reapply naming convention from explorer-merge
