+++ explorer/client/src/pages/transaction-result/TransactionResult.tsx
@@ -3,7 +3,7 @@ import { useLocation, useParams } from 'react-router-dom';
 import ErrorResult from '../../components/error-result/ErrorResult';
 import Longtext from '../../components/longtext/Longtext';
 import theme from '../../styles/theme.module.css';
-import { findDataFromID } from '../../utils/static/utility_functions';
+import { findDataFromID } from '../../utils/static/searchUtil';
 
 import styles from './TransactionResult.module.css';
