--- explorer/client/src/utils/fetch_utils.ts
@@ -1,13 +1,0 @@
-let navigateWithUnknown: Function;
-
-if (process.env.REACT_APP_DATA === 'static') {
-    import('./static/utility_functions').then(
-        (uf) => (navigateWithUnknown = uf.navigateWithUnknown)
-    );
-} else {
-    import('./internetapi/utility_functions').then(
-        (uf) => (navigateWithUnknown = uf.navigateWithUnknown)
-    );
-}
-
-export { navigateWithUnknown };
