+++ explorer/client/src/utils/internetapi/searchUtil.ts
@@ -1,6 +1,9 @@
 import { DefaultRpcClient as rpc } from './rpc';
 
-const navigateWithUnknown = async (input: string, navigate: Function) => {
+export const navigateWithUnknown = async (
+    input: string,
+    navigate: Function
+) => {
     // feels crude to just search each category for an ID, but works for now
     const addrPromise = rpc.getAddressObjects(input).then((data) => {
         if (data.length > 0) {
@@ -37,24 +40,3 @@ const navigateWithUnknown = async (input: string, navigate: Function) => {
             navigate(`../missing/${input}`);
         });
 };
-
-export function asciiFromNumberBytes(bytes: number[]) {
-    return bytes.map((n) => String.fromCharCode(n)).join('');
-}
-
-export function hexToAscii(hex: string) {
-    if (!hex || typeof hex != 'string') return;
-    hex = hex.replace(/^0x/, '');
-
-    var str = '';
-    for (var n = 0; n < hex.length; n += 2)
-        str += String.fromCharCode(parseInt(hex.substring(n, 2), 16));
-
-    return str;
-}
-
-const stdLibPrefix = /^0x2::/;
-export const trimStdLibPrefix = (str: string): string =>
-    str.replace(stdLibPrefix, '');
-
-export { navigateWithUnknown };
