--- explorer/client/src/utils/mockUtil.ts
@@ -1,6 +1,0 @@
-import mockTransactionData from './mock_data.json';
-
-export const findDataFromID = (targetID: string | undefined, state: any) =>
-    state?.category !== undefined
-        ? state
-        : mockTransactionData.data.find(({ id }) => id === targetID);
