+++ explorer/client/src/utils/searchUtil.ts
@@ -1,40 +1,13 @@
-import { DefaultRpcClient as rpc } from './SuiRpcClient';
+let navigateWithUnknown: Function;
 
-export const navigateWithUnknown = async (
-    input: string,
-    navigate: Function
-) => {
-    // TODO - replace multi-request search with backend function when ready
-    const addrPromise = rpc.getAddressObjects(input).then((data) => {
-        if (data.length <= 0) throw new Error('No objects for Address');
+if (process.env.REACT_APP_DATA === 'static') {
+    import('./static/searchUtil').then(
+        (uf) => (navigateWithUnknown = uf.navigateWithUnknown)
+    );
+} else {
+    import('./internetapi/searchUtil').then(
+        (uf) => (navigateWithUnknown = uf.navigateWithUnknown)
+    );
+}
 
-        return {
-            category: 'addresses',
-            data: data,
-        };
-    });
-
-    const objInfoPromise = rpc.getObjectInfo(input).then((data) => ({
-        category: 'objects',
-        data: data,
-    }));
-
-    //if none of the queries find a result, show missing page
-    Promise.any([objInfoPromise, addrPromise])
-        .then((pac) => {
-            if (
-                pac?.data &&
-                (pac?.category === 'objects' || pac?.category === 'addresses')
-            ) {
-                navigate(`../${pac.category}/${input}`, { state: pac.data });
-            } else {
-                throw new Error(
-                    'Something wrong with navigateWithUnknown function'
-                );
-            }
-        })
-        .catch((error) => {
-            console.log(error);
-            navigate(`../missing/${input}`);
-        });
-};
+export { navigateWithUnknown };
