+++ explorer/client/src/utils/static/searchUtil.ts
