+++ explorer/client/src/utils/stringUtils.ts
@@ -1,17 +1,18 @@
-export const asciiFromNumberBytes = (bytes: number[]): string =>
-    bytes.map((n) => String.fromCharCode(n)).join('');
+export function asciiFromNumberBytes(bytes: number[]) {
+    return bytes.map((n) => String.fromCharCode(n)).join('');
+}
 
+export function hexToAscii(hex: string) {
+    if (!hex || typeof hex != 'string') return;
+    hex = hex.replace(/^0x/, '');
+
+    var str = '';
+    for (var n = 0; n < hex.length; n += 2)
+        str += String.fromCharCode(parseInt(hex.substring(n, 2), 16));
+
+    return str;
+}
+
+const stdLibPrefix = /^0x2::/;
 export const trimStdLibPrefix = (str: string): string =>
-    str.replace(/^0x2::/, '');
-
-export const isValidHttpUrl = (url: string) => {
-    try {
-        new URL(url);
-    } catch (e) {
-        return false;
-    }
-    return /^https?/.test(url);
-};
-
-export const isSuiAddressHex = (str: string) =>
-    /^(0x)?[0-9a-fA-F]{40}$/.test(str);
+    str.replace(stdLibPrefix, '');
